requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0
httpx==0.25.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic-settings==2.0.3
